refactor: Cars 정적 팩토리 메서드를 도메인 내부에서 가능하도록 변경
