+++ src/test/java/racingcar/domain/CarsTest.java
@@ -2,6 +2,7 @@ package racingcar.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
@@ -18,12 +19,10 @@ class CarsTest {
     @DisplayName("validate() : 자동차 명이 중복이면 에러가 터진다.")
     void test_validate() {
         // given
-        String jay = "jay";
-        Car car = new Car(jay, 0);
-        List<Car> cars = List.of(car, car);
+        List<String> carNames = List.of("jay", "jay");
 
         // when & then
-        assertThatThrownBy(() -> new Cars(cars))
+        assertThatThrownBy(() -> Cars.from(carNames))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessageContaining("자동차 이름은 중복되지 않아야합니다.");
     }
@@ -35,16 +34,16 @@ class CarsTest {
         int defaultDistance = 0;
         int expectedDistanceAfterMoveSuccess = defaultDistance + 1;
 
-        Car pobi = new Car("pobi", defaultDistance);
-        Car crong = new Car("crong", defaultDistance);
-        Cars cars = new Cars(List.of(pobi, crong));
+        Cars cars = Cars.from(List.of("pobi", "crong"));
 
         // when
         cars.moveAll(randomSuccessPowerMaker);
 
         // then
-        assertThat(pobi.getDistance()).isEqualTo(expectedDistanceAfterMoveSuccess);
-        assertThat(crong.getDistance()).isEqualTo(expectedDistanceAfterMoveSuccess);
+        assertAll(
+                () -> assertThat(cars.getCars().get(0).getDistance()).isEqualTo(expectedDistanceAfterMoveSuccess),
+                () -> assertThat(cars.getCars().get(1).getDistance()).isEqualTo(expectedDistanceAfterMoveSuccess)
+        );
     }
 
     @Test
@@ -52,35 +51,31 @@ class CarsTest {
     void test_moveAll_fail() {
         // given
         int defaultDistance = 0;
-        int expectedDistanceAfterMoveSuccess = defaultDistance + 1;
-
-        Car pobi = new Car("pobi", defaultDistance);
-        Car crong = new Car("crong", defaultDistance);
-        Cars cars = new Cars(List.of(pobi, crong));
+        Cars cars = Cars.from(List.of("pobi", "crong"));
 
         // when
         cars.moveAll(randomFailPowerMaker);
 
         // then
-        assertThat(pobi.getDistance()).isNotEqualTo(expectedDistanceAfterMoveSuccess);
-        assertThat(crong.getDistance()).isNotEqualTo(expectedDistanceAfterMoveSuccess);
+        assertAll(
+                () -> assertThat(cars.getCars().get(0).getDistance()).isEqualTo(defaultDistance),
+                () -> assertThat(cars.getCars().get(1).getDistance()).isEqualTo(defaultDistance)
+        );
     }
 
     @Test
     @DisplayName("test_getWinnerNames() : 가장 많이 움직인 자동차를 반환해준다.")
     void test_getWinnerNames() {
         // given
-        Car pobi = new Car("pobi", 2);
-        Car crong = new Car("crong", 1);
-
-        Cars cars = new Cars(List.of(pobi, crong));
+        Cars cars = Cars.from(List.of("pobi", "crong"));
 
         // when
         List<String> winnerNames = cars.getWinnerNames();
 
         // then
-        assertThat(winnerNames).containsOnly(pobi.getCarName());
-        assertThat(winnerNames.size()).isEqualTo(1);
-        assertThat(winnerNames).doesNotContain(crong.getCarName());
+        assertAll(
+                () -> assertThat(winnerNames.contains("pobi")).isTrue(),
+                () -> assertThat(winnerNames.contains("crong")).isTrue()
+        );
     }
 }
